Move some common rendering logic to OpenKh.Engine
