+++ OpenKh.Engine.MonoGame/KingdomTexture.cs
@@ -1,11 +1,9 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using OpenKh.Engine.Parsers;
 using OpenKh.Kh2;
 using System;
-using System.Collections.Generic;
 
-namespace OpenKh.Game.Models
+namespace OpenKh.Engine.MonoGame
 {
     public class KingdomTexture : IDisposable
     {
@@ -30,34 +28,5 @@ namespace OpenKh.Game.Models
         {
             Texture2D?.Dispose();
         }
-    }
-
-    public class MeshGroup
-    {
-        public class Segment
-        {
-            public VertexPositionColorTexture[] Vertices { get; set; }
-        }
-
-        public class Part
-        {
-            public int[] Indices { get; set; }
-            public int TextureId { get; set; }
-            public int SegmentId { get; set; }
-            public bool IsOpaque { get; set; }
-        }
-
-        public Segment[] Segments { get; set; }
-        public Part[] Parts { get; set; }
-        public KingdomTexture[] Textures { get; set; }
-        public List<MeshDesc> MeshDescriptors { get; set; }
-    }
-
-    public class MeshDesc
-    {
-        public VertexPositionColorTexture[] Vertices;
-        public int[] Indices;
-        public int TextureIndex;
-        public bool IsOpaque;
     }
 }
