+++ OpenKh.Game/Entities/ObjectEntity.cs
@@ -1,12 +1,11 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using OpenKh.Engine.MonoGame;
 using OpenKh.Game.Debugging;
 using OpenKh.Game.Infrastructure;
-using OpenKh.Game.Models;
 using OpenKh.Kh2;
 using OpenKh.Kh2.Ard;
 using OpenKh.Kh2.Extensions;
-using System;
 using System.Linq;
 
 namespace OpenKh.Game.Entities
